Add option to disallow overwriting symbols in a SymbolTableBuilder.
By default, overwriting symbols remains allowed to stay compatible with extensions and because extensions may want to overwrite them on purpose.
However, for AsciiMath::Parser::DEFAULT_PARSER_SYMBOL_TABLE, overwriting does not make sense and is now forbidden.
This ensures that new symbols in the default parser do not overwrite existing ones, since the list is already too long to manually check this.